// src/API.ts
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type Portifolio = {
  __typename: "Portifolio",
  id: string,
  title?: string | null,
  filesKey?: Array< string | null > | null,
  backgroundImageKey?: string | null,
  content?: string | null,
  userID: string,
  startedDate?: string | null,
  endDate?: string | null,
  techologiesUsed?: Array< string | null > | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type Message = {
  __typename: "Message",
  id: string,
  text?: string | null,
  chatroomID: string,
  userID: string,
  images?: Array< string | null > | null,
  Attachments?: ModelAttachmentConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelAttachmentConnection = {
  __typename: "ModelAttachmentConnection",
  items:  Array<Attachment | null >,
  nextToken?: string | null,
};

export type Attachment = {
  __typename: "Attachment",
  id: string,
  storageKey: string,
  type: AttachmentType,
  width?: number | null,
  height?: number | null,
  duration?: number | null,
  messageID: string,
  chatroomID: string,
  createdAt: string,
  updatedAt: string,
};

export enum AttachmentType {
  IMAGE = "IMAGE",
  VIDEO = "VIDEO",
  AUDIO = "AUDIO",
}


export type ModelNotificationConnection = {
  __typename: "ModelNotificationConnection",
  items:  Array<Notification | null >,
  nextToken?: string | null,
};

export type Notification = {
  __typename: "Notification",
  id: string,
  text: string,
  userID: string,
  jobID: string,
  read?: boolean | null,
  senderID: string,
  notificationType?: NotificationType | null,
  createdAt: string,
  updatedAt: string,
};

export enum NotificationType {
  APPLIED = "APPLIED",
  COMPLETED = "COMPLETED",
  PAID = "PAID",
}


export enum UserStatus {
  REPORTED = "REPORTED",
  DELETED = "DELETED",
  FREQUENT = "FREQUENT",
  RARELY = "RARELY",
  NORMAL = "NORMAL",
}
// ...
export type UserApply = {
  __typename: "UserApply",
  id: string,
  userID?: string | null,
  jobID?: string | null,
  selectedPortifolio?: Portifolio | null,
  files?: Array< string | null > | null,
  text?: string | null,
  price?: string | null,
  userName?: string | null,
  userAvatar?: string | null,
  userStatus?: UserStatus | null,
  createdAt: string,
  updatedAt: string,
  userApplySelectedPortifolioId?: string | null,
};
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type CreateUserApplyInput = {
  id?: string | null,
  userID?: string | null,
  jobID?: string | null,
  files?: Array< string | null > | null,
  text?: string | null,
  price?: string | null,
  userName?: string | null,
  userAvatar?: string | null,
  userStatus?: UserStatus | null,
  userApplySelectedPortifolioId?: string | null,
};

export type ModelUserApplyConditionInput = {
  userID?: ModelIDInput | null,
  jobID?: ModelIDInput | null,
  files?: ModelStringInput | null,
  text?: ModelStringInput | null,
  price?: ModelStringInput | null,
  userName?: ModelStringInput | null,
  userAvatar?: ModelStringInput | null,
  userStatus?: ModelUserStatusInput | null,
  and?: Array< ModelUserApplyConditionInput | null > | null,
  or?: Array< ModelUserApplyConditionInput | null > | null,
  not?: ModelUserApplyConditionInput | null,
  userApplySelectedPortifolioId?: ModelIDInput | null,
};

export type ModelUserStatusInput = {
  eq?: UserStatus | null,
  ne?: UserStatus | null,
};

export type UpdateUserApplyInput = {
  id: string,
  userID?: string | null,
  jobID?: string | null,
  files?: Array< string | null > | null,
  text?: string | null,
  price?: string | null,
  userName?: string | null,
  userAvatar?: string | null,
  userStatus?: UserStatus | null,
  userApplySelectedPortifolioId?: string | null,
};

export type DeleteUserApplyInput = {
  id: string,
};

export type CreateNotificationInput = {
  id?: string | null,
  text: string,
  userID: string,
  jobID: string,
  read?: boolean | null,
  senderID: string,
  notificationType?: NotificationType | null,
};

export type ModelNotificationConditionInput = {
  text?: ModelStringInput | null,
  userID?: ModelIDInput | null,
  jobID?: ModelIDInput | null,
  read?: ModelBooleanInput | null,
  senderID?: ModelStringInput | null,
  notificationType?: ModelNotificationTypeInput | null,
  and?: Array< ModelNotificationConditionInput | null > | null,
  or?: Array< ModelNotificationConditionInput | null > | null,
  not?: ModelNotificationConditionInput | null,
};

export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelNotificationTypeInput = {
  eq?: NotificationType | null,
  ne?: NotificationType | null,
};
// ...
export type ModelUserConditionInput = {
  fullname?: ModelStringInput | null,
  reviewsLength?: ModelIntInput | null,
  totalJobsDone?: ModelIntInput | null,
  email?: ModelStringInput | null,
  phoneNumber?: ModelStringInput | null,
  avatar?: ModelStringInput | null,
  notificationsActived?: ModelBooleanInput | null,
  website?: ModelStringInput | null,
  firstTimeLoggin?: ModelBooleanInput | null,
  categories?: ModelStringInput | null,
  hasDeficiency?: ModelBooleanInput | null,
  points?: ModelIntInput | null,
  profission?: ModelStringInput | null,
  backgroundImage?: ModelStringInput | null,
  status?: ModelUserStatusInput | null,
  about?: ModelStringInput | null,
  cv?: ModelStringInput | null,
  and?: Array< ModelUserConditionInput | null > | null,
  or?: Array< ModelUserConditionInput | null > | null,
  not?: ModelUserConditionInput | null,
  userSelectedPortifolioId?: ModelIDInput | null,
};

export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type UpdateUserInput = {
  id: string,
  fullname?: string | null,
  reviewsLength?: number | null,
  totalJobsDone?: number | null,
  email?: string | null,
  phoneNumber?: string | null,
  avatar?: string | null,
  notificationsActived?: boolean | null,
  website?: string | null,
  firstTimeLoggin?: boolean | null,
  categories?: Array< string | null > | null,
  hasDeficiency?: boolean | null,
  points?: number | null,
  profission?: string | null,
  backgroundImage?: string | null,
  status?: UserStatus | null,
  about?: string | null,
  cv?: string | null,
  userSelectedPortifolioId?: string | null,
};
// ...
export type CreateChatRoomInput = {
  id?: string | null,
  name?: string | null,
  chatRoomLastMessageId?: string | null,
};

export type ModelChatRoomConditionInput = {
  name?: ModelStringInput | null,
  and?: Array< ModelChatRoomConditionInput | null > | null,
  or?: Array< ModelChatRoomConditionInput | null > | null,
  not?: ModelChatRoomConditionInput | null,
  chatRoomLastMessageId?: ModelIDInput | null,
};

export type UpdateChatRoomInput = {
  id: string,
  name?: string | null,
  chatRoomLastMessageId?: string | null,
};

export type DeleteChatRoomInput = {
  id: string,
};

export type CreateMessageInput = {
  id?: string | null,
  text?: string | null,
  chatroomID: string,
  userID: string,
  images?: Array< string | null > | null,
};

export type ModelMessageConditionInput = {
  text?: ModelStringInput | null,
  chatroomID?: ModelIDInput | null,
  userID?: ModelIDInput | null,
  images?: ModelStringInput | null,
  and?: Array< ModelMessageConditionInput | null > | null,
  or?: Array< ModelMessageConditionInput | null > | null,
  not?: ModelMessageConditionInput | null,
};
// ...
export type CreateUserChatRoomInput = {
  id?: string | null,
  userID: string,
  chatRoomID: string,
};

export type ModelUserChatRoomConditionInput = {
  userID?: ModelIDInput | null,
  chatRoomID?: ModelIDInput | null,
  and?: Array< ModelUserChatRoomConditionInput | null > | null,
  or?: Array< ModelUserChatRoomConditionInput | null > | null,
  not?: ModelUserChatRoomConditionInput | null,
};
// ...
export type CreateUserApplyMutationVariables = {
  input: CreateUserApplyInput,
  condition?: ModelUserApplyConditionInput | null,
};

export type CreateUserApplyMutation = {
  createUserApply?:  {
    __typename: "UserApply",
    id: string,
    userID?: string | null,
    jobID?: string | null,
    selectedPortifolio?:  {
      __typename: "Portifolio",
      id: string,
      title?: string | null,
      filesKey?: Array< string | null > | null,
      backgroundImageKey?: string | null,
      content?: string | null,
      userID: string,
      startedDate?: string | null,
      endDate?: string | null,
      techologiesUsed?: Array< string | null > | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    files?: Array< string | null > | null,
    text?: string | null,
    price?: string | null,
    userName?: string | null,
    userAvatar?: string | null,
    userStatus?: UserStatus | null,
    createdAt: string,
    updatedAt: string,
    userApplySelectedPortifolioId?: string | null,
  } | null,
};
// ...
export type CreateNotificationMutationVariables = {
  input: CreateNotificationInput,
  condition?: ModelNotificationConditionInput | null,
};

export type CreateNotificationMutation = {
  createNotification?:  {
    __typename: "Notification",
    id: string,
    text: string,
    userID: string,
    jobID: string,
    read?: boolean | null,
    senderID: string,
    notificationType?: NotificationType | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type UpdateUserMutationVariables = {
  input: UpdateUserInput,
  condition?: ModelUserConditionInput | null,
};
// ...
export type CreateChatRoomMutationVariables = {
  input: CreateChatRoomInput,
  condition?: ModelChatRoomConditionInput | null,
};
// ...
export type UpdateChatRoomMutationVariables = {
  input: UpdateChatRoomInput,
  condition?: ModelChatRoomConditionInput | null,
};
// ...
export type CreateMessageMutationVariables = {
  input: CreateMessageInput,
  condition?: ModelMessageConditionInput | null,
};

export type CreateMessageMutation = {
  createMessage?:  {
    __typename: "Message",
    id: string,
    text?: string | null,
    chatroomID: string,
    userID: string,
    images?: Array< string | null > | null,
    Attachments?:  {
      __typename: "ModelAttachmentConnection",
      items:  Array< {
        __typename: "Attachment",
        id: string,
        storageKey: string,
        type: AttachmentType,
        width?: number | null,
        height?: number | null,
        duration?: number | null,
        messageID: string,
        chatroomID: string,
        createdAt: string,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type CreateUserChatRoomMutationVariables = {
  input: CreateUserChatRoomInput,
  condition?: ModelUserChatRoomConditionInput | null,
};
// ...
export type GetPortifolioQueryVariables = {
  id: string,
};

export type GetPortifolioQuery = {
  getPortifolio?:  {
    __typename: "Portifolio",
    id: string,
    title?: string | null,
    filesKey?: Array< string | null > | null,
    backgroundImageKey?: string | null,
    content?: string | null,
    userID: string,
    startedDate?: string | null,
    endDate?: string | null,
    techologiesUsed?: Array< string | null > | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};
// ...
export type GetChatRoomQueryVariables = {
  id: string,
};